Split Document and Webpage contexts
Both needed to be split at once in order to avoid a circular dependency.
